+++ src/ops/padToSquare.ts
@@ -42,7 +42,9 @@ export function padToSquare(
       paddingTensorAppend
     ]
       .filter(t => !!t)
-      .map((t: tf.Tensor) => t.toFloat()) as tf.Tensor4D[]
+      // .map((t: tf.Tensor) => t.toFloat()) as tf.Tensor4D[]
+      .map((t: tf.Tensor) => tf.cast(t, 'float32')) as tf.Tensor4D[]
     return tf.concat(tensorsToStack, paddingAxis)
+
   })
 }
